rename model classes with suffix that fit them in their respective modules to easily read the code (eg a model class supposed to be an entity to end with suffix entity)
