+++ domain/src/main/java/com/aliasadi/domain/usecase/GetFavoriteMovies.kt
@@ -1,6 +1,6 @@
 package com.aliasadi.domain.usecase
 
-import com.aliasadi.domain.entities.MovieEntity
+import com.aliasadi.domain.models.MovieModel
 import com.aliasadi.domain.repository.MovieRepository
 import com.aliasadi.domain.util.Result
 
@@ -10,5 +10,5 @@ import com.aliasadi.domain.util.Result
 class GetFavoriteMovies(
     private val movieRepository: MovieRepository
 ) {
-    suspend fun getFavoriteMovies(): Result<List<MovieEntity>> = movieRepository.getFavoriteMovies()
+    suspend fun getFavoriteMovies(): Result<List<MovieModel>> = movieRepository.getFavoriteMovies()
 }
